Adding zurb foundation
Adding zurb foundation to help with UI layouts

+++ app/models/funding_requirement.rb
@@ -1,7 +1,7 @@
 class FundingRequirement
   include Mongoid::Document
   field :desired_amount, type: BigDecimal
-  field :percentage_to_give, type: BigDecimal
+  field :percentage_to_give, type: Float
 
   embedded_in :company
 end
